% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMixtureModel.R
\name{GaussianMixtureModel}
\alias{GaussianMixtureModel}
\title{Gaussian Mixture Model}
\usage{
GaussianMixtureModel(X, k, initial_mu, max_iter = 100, tol = 1e-08)
}
\arguments{
\item{X}{an N by p data matrix}

\item{k}{integer, number of Gaussian components}

\item{initial_mu}{a k by p matrix, the initial points of the EM Algorithm; each row corresponds to one component}

\item{max_iter}{integer, the maximum number of EM iterations}

\item{tol}{scaler or NULL, used as the convergence criteria \cr
if tol is scaler, if the distance between centroids in the current iteration and the previous iteration is less than tol, then the algorithm stops early \cr
if tol is NULL, then early stopping criterion does not apply}
}
\value{
A list containing mu, Sigma, r and loglikelihood \cr
mu: a k by p matrix, fitted centroids; each row corresponds to one component \cr
Sigma: a p by p by k array, fitted covariance matrices \cr
r: an N by k matrix, fitted responsibility; each row corresponds to one sample; the value (i, j) is the fitted probability of sample i belonging to component j \cr
}
\description{
Fit a Gaussian Mixture Model by the EM Algorithm
}
\examples{
# create samples from a two-component Gaussian Mixture
X = matrix(0, nrow = 500, ncol = 2)
z = sample(2, 500, replace = TRUE)
X[which(z == 1), ] = rnorm(sum(z == 1) * 2, mean = 0, sd = 1)
X[which(z == 2), ] = rnorm(sum(z == 2) * 2, mean = 5, sd = 1)

# fit a Gaussian Mixture Model to the data
gmm = GaussianMixtureModel(X, 2, initial_mu = matrix(c(0, 1, 0, 1), nrow = 2, ncol = 2))

# prediction
centroids = gmm$mu
r = gmm$r
cluster = apply(r, 1, which.max)

# plot
plot(X[, 1], X[, 2], col = c('orange', 'red')[cluster], cex = 0.5)
points(x = centroids[, 1], y = centroids[, 2], pch = 10, col = 'blue', cex = 2)
legend(x = 'topleft',
       legend = c('cluster 1', 'cluster 2', 'centers'),
       col = c('orange', 'red', 'blue'),
       pch = c(1, 1, 10))

}
